+++ app/src/main/java/com/capyreader/app/ui/settings/GeneralSettingsViewModel.kt
@@ -1,7 +1,6 @@
 package com.capyreader.app.ui.settings
 
 import android.app.Application
-import android.content.Context
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
@@ -10,12 +9,9 @@ import com.capyreader.app.common.AppPreferences
 import com.capyreader.app.refresher.RefreshInterval
 import com.capyreader.app.refresher.RefreshScheduler
 import com.jocmp.capy.Account
-import com.jocmp.capy.AccountManager
 import com.jocmp.capy.accounts.AutoDelete
-import com.jocmp.capy.accounts.Source
 
 class GeneralSettingsViewModel(
-    private val accountManager: AccountManager,
     private val refreshScheduler: RefreshScheduler,
     val account: Account,
     private val appPreferences: AppPreferences,
@@ -29,11 +25,6 @@ class GeneralSettingsViewModel(
 
     var canOpenLinksInternally by mutableStateOf(appPreferences.openLinksInternally.get())
         private set
-
-    val accountSource: Source = account.source
-
-    val accountName: String
-        get() = account.preferences.username.get()
 
     fun updateRefreshInterval(interval: RefreshInterval) {
         refreshScheduler.update(interval)
@@ -56,12 +47,4 @@ class GeneralSettingsViewModel(
     fun clearAllArticles() {
         account.clearAllArticles()
     }
-
-    fun removeAccount() {
-        appPreferences.clearAll()
-        accountManager.removeAccount(accountID = account.id)
-    }
-
-    private val applicationContext: Context
-        get() = getApplication<Application>().applicationContext
 }
