+++ app/src/main/java/com/capyreader/app/ui/settings/SettingsModule.kt
@@ -17,9 +17,16 @@ val settingsModule = module {
     }
     viewModel {
         GeneralSettingsViewModel(
+            refreshScheduler = get(),
+            account = get(),
+            appPreferences = get(),
+            application = get()
+        )
+    }
+    viewModel {
+        AccountSettingsViewModel(
             account = get(),
             accountManager = get(),
-            refreshScheduler = get(),
             appPreferences = get(),
             application = get()
         )
