+++ app/src/main/res/values/strings.xml
@@ -136,7 +136,7 @@
     <string name="feed_list_top_bar_close_search">Dismiss search</string>
     <string name="settings_panel_general_title">General</string>
     <string name="settings_panel_display_title">Display &amp; Appearance</string>
-    <string name="settings_import_export_title">Import &amp; Export</string>
+    <string name="settings_account_title">Account</string>
     <string name="settings_about_title">About</string>
     <string name="settings_list_top_bar_title">Settings</string>
     <string name="settings_auto_delete_articles_title">Keep articles</string>
