+++ capy/src/test/java/com/jocmp/capy/persistence/ArticleMapperTest.kt
@@ -23,6 +23,7 @@ class ArticleMapperTest {
             read = false,
             starred = false,
             feedTitle = "",
+            feedURL = null,
             faviconURL = null
         )
